NeMo/Code/webserver/requirements.python.txt
dash==2.8.0
dash_daq==0.5.0
git+https://github.com/openskynetwork/opensky-api.git#egg=opensky-api&subdirectory=python
python-dotenv==0.21.0
dash-leaflet==0.1.23
pytz

chardet
nemo
Cython
nemo_toolkit[all]